mi vere forgesis kion mi faris, nur la de la kolicioj de la ludantoj.
mi dauxris kiel tri tagoj farinte unu maniero por la kolicio sed nun mi
havas alian manieron por tio plie simpla

+++ Projekto n01 - Tankoj la atakoj/src/qef/statmayin/statj/lud/Ludperant.java
@@ -40,7 +40,7 @@ public class Ludperant implements Statlud {
 		for(int i = 0; i<Vicperant.ludantj.length; i++) {
 			DebugDesegn.desegnString(Vicperant.ludantj[i].nomn(), Konstantj.ludLargx - 150, 78 + 14*i,
 					Vicperant.ludantj[i].kolorn());
-			DebugDesegn.desegnString(Vicperant.ludantj[i].monn() + "", Konstantj.ludLargx - 60, 78 + 14*i);
+			DebugDesegn.desegnString((int) Vicperant.ludantj[i].monn() + "", Konstantj.ludLargx - 60, 78 + 14*i);
 		}
 	}
